massa_versioning

Module versioning

Structs

Active

State machine for a Versioning component that tracks the deployment state

Advance

A message to update the ComponentState

AdvanceLW

Defined

State machine for a Versioning component that tracks the deployment state

Failed

State machine for a Versioning component that tracks the deployment state

LockedIn

State machine for a Versioning component that tracks the deployment state

MipInfo

MIP info (name & versions & time range for a MIP)

MipState

Wrapper of ComponentState (in order to keep state history)

MipStatsConfig

Statistics in MipStoreRaw

MipStore

Database for all MIP info

MipStoreRaw

Store of all versioning info

MipStoreStats 🔒

In order for a MIP to be accepted, we compute statistics about other node ‘network’ version announcement

Started

State machine for a Versioning component that tracks the deployment state

Enums

ComponentState

State machine for a Versioning component that tracks the deployment state

ComponentStateMessages

ComponentStateTypeId

ExtendFromDbError

Error returned by ’extend_from_db`

IsConsistentError

Error returned by MipState::is_consistent_with

IsConsistentWithShutdownPeriodError

Error returned by ’is_consistent_with_shutdown_period`

IsKVValidError

MipComponent

Versioning component enum

StateAtError

Error returned by MipStateHistory::state_at

UpdateWithError

Error returned by MipStoreRaw::update_with