massa_versioning

Module versioning_factory

Enums

FactoryError

Factory error

FactoryStrategy

Strategy to use when creating a new object from a factory

Traits

VersioningFactory

Trait for Factory that create objects based on versioning store